Minor cleanup (makes icc happier).
git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/tiptop/trunk@80 fd2ea14d-3e48-4a98-bac5-13ccbbe52f7f

+++ tiptop.c
@@ -248,7 +248,7 @@ static void batch_mode(struct process_list* proc_list, screen_t* screen)
   foo = system("date");
   foo = system("uname -a");
   printf("delay: %.2f  idle: %d  threads: %d\n",
-         options.delay, options.idle, options.show_threads);
+         options.delay, (int)options.idle, (int)options.show_threads);
   if (options.watch_pid) {
     printf("watching pid %d\n", options.watch_pid);
   }
